+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes of releases are documented in this file
 using the [Keep a CHANGELOG](https://keepachangelog.com/) principles.
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed binary detection to also support umlauts
+
+
 ## [1.20.0]
 
 ### Added